# refactor(makeStyles): add "ax" to perform concatenation of atomic classess