--- a/packages/make-styles/src/constants.ts
+++ b/packages/make-styles/src/constants.ts
@@ -1,8 +1,6 @@
-import { MakeStylesLookupEntry } from './types';
-
 export const CAN_USE_CSS_VARIABLES = window.CSS && CSS.supports('color', 'var(--c)');
 
 export const HASH_PREFIX = 'f';
 export const RTL_PREFIX = 'r';
 
-export const DEFINITION_LOOKUP_TABLE: Record<string, MakeStylesLookupEntry> = {};
+export const DEFINITION_LOOKUP_TABLE: Record<string, string> = {};