--- a/packages/make-styles/src/index.ts
+++ b/packages/make-styles/src/index.ts
@@ -1,4 +1,6 @@
 export { createDOMRenderer } from './renderer/createDOMRenderer';
 
+export { ax } from './ax';
 export { makeStyles } from './makeStyles';
+
 export * from './types';