--- a/packages/make-styles/src/makeStyles.ts
+++ b/packages/make-styles/src/makeStyles.ts
@@ -1,4 +1,4 @@
-import { CAN_USE_CSS_VARIABLES, DEFINITION_LOOKUP_TABLE } from './constants';
+import { CAN_USE_CSS_VARIABLES } from './constants';
 import { createCSSVariablesProxy, resolveDefinitions } from './runtime/index';
 import {
   MakeStylesDefinition,
@@ -13,11 +13,6 @@ export function makeStyles<Selectors, Tokens>(
 ) {
   const cxCache: Record<string, string> = {};
 
-  function computeClasses(
-    selectors: Selectors,
-    options: MakeStylesOptions<Tokens>,
-    ...classNames: (string | undefined)[]
-  ): string;
   function computeClasses(selectors: Selectors, options: MakeStylesOptions<Tokens>): string {
     let tokens: Tokens | null;
     let resolvedDefinitions: MakeStylesResolvedDefinition<Selectors, Tokens>[];
@@ -41,41 +36,6 @@ export function makeStyles<Selectors, Tokens>(
       );
     }
 
-    let nonMakeClasses: string = '';
-    const overrides: MakeStylesMatchedDefinitions = {};
-    let overridesCx: string = '';
-
-    // arguments are parsed intentionally manually to avoid double loops as TS transforms rest via an additional loop
-    for (let i = 2; i < arguments.length; i++) {
-      // eslint-disable-next-line prefer-rest-params
-      const classNames = arguments[i];
-
-      if (typeof classNames === 'string' && classNames !== '') {
-        // .split() is an expensive call, it's faster to ensure that string contains any spaces before splitting
-        if (classNames.indexOf(' ') === -1) {
-          const definition = DEFINITION_LOOKUP_TABLE[classNames];
-
-          if (definition !== undefined) {
-            overrides[definition[0]] = DEFINITION_LOOKUP_TABLE[classNames][1];
-            overridesCx += classNames;
-          } else {
-            nonMakeClasses += classNames + ' ';
-          }
-        } else {
-          classNames.split(' ').forEach(className => {
-            const definition = DEFINITION_LOOKUP_TABLE[className];
-
-            if (definition !== undefined) {
-              overrides[definition[0]] = definition[1];
-              overridesCx += className;
-            } else {
-              nonMakeClasses += className + ' ';
-            }
-          });
-        }
-      }
-    }
-
     let matchedIndexes = '';
     const matchedDefinitions: MakeStylesMatchedDefinitions[] = [];
 
@@ -88,22 +48,21 @@ export function makeStyles<Selectors, Tokens>(
       }
     }
 
-    const overridesHash = overridesCx === '' ? '' : overridesCx;
-    const cxCacheKey = options.renderer.id + matchedIndexes + '' + overridesHash;
+    const cxCacheKey = options.renderer.id + matchedIndexes;
     const cxCacheElement = cxCache[cxCacheKey];
 
     if (CAN_USE_CSS_VARIABLES && cxCacheElement !== undefined) {
-      return nonMakeClasses + cxCacheElement;
+      return cxCacheElement;
     }
 
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
-    const resultDefinitions: MakeStylesMatchedDefinitions = Object.assign({}, ...matchedDefinitions, overrides);
-    const resultClasses = options.renderer.insertDefinitions(DEFINITION_LOOKUP_TABLE, resultDefinitions, !!options.rtl);
+    const resultDefinitions: MakeStylesMatchedDefinitions = Object.assign({}, ...matchedDefinitions);
+    const resultClasses = options.renderer.insertDefinitions(resultDefinitions, !!options.rtl);
 
     cxCache[cxCacheKey] = resultClasses;
 
-    return nonMakeClasses + resultClasses;
+    return resultClasses;
   }
 
   return computeClasses;