--- a/packages/make-styles/src/runtime/utils/hashString.ts
+++ b/packages/make-styles/src/runtime/utils/hashString.ts
@@ -1,0 +1,7 @@
+import murmurHash from '@emotion/hash';
+
+export function hashString(value: string): string {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  return murmurHash(value).padEnd(7, '0');
+}