--- a/packages/make-styles/src/runtime/utils/hashString.ts
+++ b/packages/make-styles/src/runtime/utils/hashString.ts
@@ -1,7 +1,22 @@
 import murmurHash from '@emotion/hash';
 
+function padEndHash(value: string): string {
+  const hashLength = value.length;
+
+  if (hashLength === 7) {
+    return value;
+  }
+
+  for (let i = hashLength; i < 7; i++) {
+    value += '0';
+  }
+
+  return value;
+}
+
+/**
+ * Uses MurmurHash to generate a hash and ensures that its output is always has 7 characters.
+ */
 export function hashString(value: string): string {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  return murmurHash(value).padEnd(7, '0');
+  return padEndHash(murmurHash(value));
 }