## 🔧 `ax()` function to merge classes

`ax()` functions performs concatenation and deduplication of atomic classnames. Previously we used `makeStyles` util for this, but it's better to separate responsibilities:

#### Before

```ts
const overrides = useOverrides({});
const classes = useStyles({}, overrides);
```

#### After

```ts
const classes = ax(useStyles({}), useOverrides({}));
```

## 🏭 Implementation changes

Overall, the idea is inspired by [ax()](https://github.com/atlassian-labs/compiled/blob/82be1a394ff5b70b2bcc29706b603140ef759f0c/packages/react/src/runtime/ax.tsx#L30) implementation in [Compiled](https://github.com/atlassian-labs/compiled). However, there is a significant difference in approaches.

I decided to go with an idea of sequence hashes, this allows make a path for lookups significantly shorter. It has also linear improvements based on amount of classes.

As a simplest example we can take these styles:

```js
{ color: 'red', margin: '5px', position: 'absolute' }
```

#### Before

```tsx
// 💡 output of `makeStyles()` call, we need to lookup to map 6 classes to properties (color, position and 4 for margins)
fe3e8s90 f1euv43f f1rqyxcv fq02s400 f475ppk0 f1f7bkv5
```

#### After

```tsx
// "__3lptt90" is a utility classname with zero styles, it's hash of an atomic set of classes
// this allows to perform a **single** lookup to map properties
__3lptt90 fe3e8s90 f1euv43f f1rqyxcv fq02s400 f475ppk0 f1f7bkv5
```

## 🏎️ Performance?

| Case                           |                           Result |
| ------------------------------ | -------------------------------: |
| Previous impl. (small set)     |                   827068 ops/sec |
| Previous impl. (long set)      |                   205659 ops/sec |
| compliled approach (small set) |                   628163 ops/sec |
| compliled approach (long set)  |                   145842 ops/sec |
| `ax()` (small set)             |                  1528328 ops/sec |
| `ax()` (long set)              | 3245282 ops/sec (**15x faster**) |

- small set was represented by 2 classes
- long set was represented by 25 classes